feat: 🎸 TextTyper
Render a list of characters, on keyDown check for a backspace and put
the current index status to not typed (the current index does not hcange
if it's currently 0 or in an ERROR status). If a capslock or shift is
fired, ignore them, otherwise if the typed key does not correspond to
the character at the current index in the text prop, put the current
index status in error mode, otherwise, advance the current index from 1.