feat: 🎸 Enable wpm calculation in the component
Add a start and end timestamp in the state. When both a correct/errored
key is pressed, in the case where the index is 0, we set the start
timestamp. When the current index is equal to the last index of the
text, we set the end timestamp and the wpm is calculated from those
variables.